15-30

+++ src/lab2/task15.java
@@ -5,15 +5,15 @@ import java.util.Scanner;
 public class task15 {
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
-        int temperatureInC;
-        temperatureInC = s.nextInt();
+        double temperatureInC;
+        temperatureInC = s.nextDouble();
         if (temperatureInC < -273.15){
             System.out.println("Температура должна быть выше абсолютного нуля");
             return;
         }
         double temperatureInF;
         temperatureInF = temperatureInC *1.8+32;
-        System.out.printf("%.5f",temperatureInF);
+        System.out.printf("%.4f",temperatureInF);
 
 
     }
